Lift out the formatting into a separate function.
The formatting itself is now lifted out to _cee_vformat(), which
returns a const char *, and puts the JSON object into its first
(output) parameter.

We then export cee_format() and cee_vformat(), that use this, strdup
and return the result, and free the JSON object too.

cee_vsyslog() was changed to use _cee_vformat(), because it does not
need to strdup() the result.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>