Implement the new syslog() overrides.
Changed the API, so that we have cee_(v)syslog(), which is the
enhanced syslog API, that takes key-value pairs in addition to the
traditional message, and outputs CEE-JSON to syslog.

We also have (v)syslog() overrides, that do the same, just without the
extra key-value pairs.

Both of these add some extra stuff, via _cee_discover(): at the
moment, this is the facility, the priority and the pid.

We also override openlog(), that stores some info in thread-local
storage. That info is used by _cee_discover().

For future use, this patch also defines LOG_CEE_NODISCOVER and
LOG_CEE_NOCACHE. They're not used anywhere yet, though.

Signed-off-by: Gergely Nagy <algernon@madhouse-project.org>