Implement wrappers for __syslog_chk, when appropriate.
When compiled with __USE_FORTIFY_LEVEL, software using <syslog.h> will
use the __syslog_chk function instead of syslog(), so we need to wrap
that too.

To do this properly, the library will wrap __syslog_chk when built
with fortification, and use that from both the syslog() and the
__syslog_chk() overrides. If built without, it will use the standard
syslog() function.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>