Fix infinite recursion on FreeBSD.
On FreeBSD, syslog() calls vsyslog(), our own replacement version. So
when from syslog(), we call our vsyslog(), which calls the old
syslog(), that ends up calling our own vsyslog() again.

To work around this issue, use a thread-local static variable that
signals whether we're recursing, and if we are, then call the old
vsyslog from the new one, thus breaking the loop.

This makes the library function correctly on FreeBSD too.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>