Properly handle va_lists, by copying.
Up until now, the library abused that passing va_list around without
copying works on various architectures. On others, that does not.

To remedy the issue, the library now parses format strings itself, and
copies va_list. This makes it work on all architectures, independent
on the internal implementation of va_list, and preserves both API and
ABI.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>