Format JSON ourselves, without json-c.
To make the library thinner, we now format JSON ourselves: there was
so little we used from json-c, that doing it ourselves is lighter and
faster.

With this commit, we have the architecture in place, with only a few
little things remaining to be done, such as escaping string values.

Signed-off-by: Gergely Nagy <algernon@balabit.hu>