Properly reset the JSON buffer in all cases.
Previously the JSON buffer was only reset when ul_vformat() was
called, but not everything goes through that, so some messages ended
up being garbled.

The proper thing to do is to reset in _ul_vformat(), as everything
goes through that, and that's the right place to do this anyway.

Reported-by: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>