Resolve facility & priority properly
The priority passed to the syslog() call can contain a facility
embedded, so we need to extract it to find both the facility and the
priority.

This also means that the facility set at openlog()-time can be
overridden later, so _find_facility() has to take the priority value
into account.

This patch makes libumberlog do just the above: extract both priority
and facility from the priority passed to syslog(), and if facility is
0, use the default set at openlog()-time, if any.

Added a test case to test this expected behaviour.

Reported-by: Peter Czanik <czanik@balabit.hu>
Signed-off-by: Gergely Nagy <algernon@balabit.hu>