Let libc manage the syslog mask.
glibc has a per-process, not per-thread, syslog mask, so the per-thread
mask in ul_sys_settings may end up desynchronized from the glibc mask,
breaking compatibility for LD_PRELOAD.

Just call setlogmask (0) to always read the glibc mask, and drop the
per-thread cache.

An alternative is to use a per-thread mask for ul_* calls and to use the
glibc mask for the LD_PRELOAD-overridden functions, but that seems to be
unnecessarily complex.

ul_setlogmask () is now a trivial wrapper, so we don't need to override
setlogmask () at all.  We could also drop ul_setlogmask () completely,
but that would break the API.

Signed-off-by: Miloslav Trmač <mitr@redhat.com>